fix pydocstyle
